Prevent non user from editing a pin

+++ routes/deletePin.js
@@ -6,19 +6,25 @@ const db = require('../lib/pinQueries');
 
 router.post("/", (req, res) => {
   const pin_id = req.body.pinId;
+  const user_id = req.session.userId;
   const map = {
     pin_id
   };
 
-  db.deletePinById(map)
-    .then(data => {
-      res.redirect(`/maps/${data[0].map_id}`);
-    })
-    .catch(err => {
-      res
-        .status(500)
-        .json({ error: err.message });
-    });
+  if (!user_id) {
+    return;
+  } else {
+    db.deletePinById(map)
+      .then(data => {
+        res.redirect(`/maps/${data[0].map_id}`);
+      })
+      .catch(err => {
+        res
+          .status(500)
+          .json({ error: err.message });
+      });
+  }
+
 });
 
 module.exports = router;
